style: renamed and repositioned files
